requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
pypdf>=3.17.0
pymilvus>=2.3.0
openai>=1.0.0
python-dotenv>=1.0.0

